migrate over wiki.willnorris.com content
that site was never really a wiki... really just notes to myself.  For
now I'm moving them to willnorris.com/wiki, but they really need to find
a more permanent home.  The main rationle for moving them off of
wiki.willnorris.com was to ease management, and to get everything on
HTTPS.