update display of post timestamp
this is really hacky, but the only way I can figure out for now to
properly display the original timezone for when the post was written.
Working with timezones is a complete pain, but this hack works for now
since all my posts are in the Central and Pacific time zones.