support current and future syndication links
links with a path are treated as existing syndicated links and are
made available in the posts.syndication_links array.  links with no path
are treated as posse links. posts.bridgy_links will return the brid.gy
targets for sites that it supports (currently Twitter and Facebook) for
those links.