make _config.yml my development config
jekyll 3.1 broke the ability to override an existing hash value with
null (jekyll/jekyll#4494), which is what I was doing in _debug.yml.  The
simple solution is to move my production overrides into _production.yml
and put the development values in _config.yml.  Which is actually more
consistent with how we often do things at work anyway (always default to
non-production values)