nginx: remove extra redirect for go subpackages
As documented at https://wjn.me/b/4_m1, requests for
/go/imageproxy/cmd/imageproxy?go=get=1 are rewritten to /go/imageproxy.
However, because that is a directory on disk, nginx returns a 301
redirect to /go/imageproxy/ (with a trailing slash).  The go command
actually handles that redirect fine, but we can easily eliminate the
need for it by including the trailing slash ourselves in the rewrite.