modify store-webmention.js for my use
update domain and storage location.  Also flatten path of target URL for
storing webmentions (so "2020:01:post-title" rather than
"2020/01/post-title").  This is to work around apparent limitations with
Hugo and go templates, which is kind of a bummer, but not the end of the
world.

rewrite all target URLs to HTTPS, since some old ones still use HTTP,
which messes up webmention file paths.