extract imageproxy-url partial from img shortcode
This allows imageproxy URLs to be created from template files (or
partials) in addition to content pages that use the shortcode. Use this
new partial to proxy all comment author images.

This partial includes the ability to base64 encode the remote URL, which
isn't actually supported by imageproxy yet. I have it working in a
local copy, which I'll probably work on cleaning up and committing.
Otherwise, I'll take this option out later.