static/favicon.svg: update with logo from blog post
replace favicon.svg with the version from my https://wjn.me/2023/logo
blog post. The logo in the blog post was hand-edited based on an
original triangle 60px on each side.  The gap on the left side is
exactly 6 px. I think this other logo has all the same dimensions,
though the gap may be a little off.  So, replace with the one I know is
right.