webmentions: render any mention with content
I noticed that a number of webmentions are typed as a "like" or
"bookmarK", but also have content. So show any mentions that include
content.  Except for reposts, since the content is typically just the
original post.  Sometimes reposts can have original content also, but I
don't have a great way to identify that.

Also render mentions in individual light gray boxes so that they are
visually distinct from the main article, and space them out a tad.