css: use postcss in place of sass
I've run into some annoying issues with the go implementation of sass,
and trying to switch to the Dart version is not worth it.  Besides, the
interesting features I actually use(d) in sass, variables and nesting,
are now part of modern CSS, so there's really no need for sass.

For now, postcss is still expanding nested styles, similar to sass, but
I'll probably drop that before long.